+++ scenes/block.scn
@@ -81,11 +81,13 @@
     },
     {
       "position": [-119, 1.15, -359],
-      "start_url": "https://webaverse.github.io/pistol/"
+      "start_url": "https://webaverse.github.io/rocket-launcher/"
     },
     {
-      "position": [5, 1, 20],
-      "start_url": "/Assets/ball"
+      "position": [
+        -130, 1.15, -359
+      ],
+      "start_url": "https://webaverse.github.io/ball/"
     },
     {
       "position": [
