+++ src/components/general/crosshair/crosshair.module.css
@@ -1,10 +1,12 @@
-
 .crosshair {
     position: fixed;
     top: 0;
     bottom: 0;
     left: 0;
     right: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
     z-index: 1;
     user-select: none;
     pointer-events: none;
